backend/requirements.txt
Flask
Flask-SQLAlchemy
flask-cors
google-generativeai
python-docx
PyMuPDF
Werkzeug
